requirements.txt
streamlit
python-dotenv
urllib3
langchain
langchain-core
langchain-community
langchain-chroma
langchain-groq
langchain-huggingface
chromadb
unstructured
sentence-transformers
tiktoken
